Found a bug in "Update Bug"
I found a bug in "Update Bug" menu. If you enter a bug ID that does not
exist, it creates a null pointer exception. I added in some exception
handling so that it doesn’t crash. It now just shows a blank screen.
Next step is to create screen that shows a message that the Bug ID
entered does not exist. I tried for sometime to do that, but couldn’t
get it to work.